test(api): update and add test on datagouv parser
